Moved docker code out of craft package.
