+++ cmd/backup.go
@@ -3,6 +3,8 @@ package cmd
 import (
 	"log"
 
+	"github.com/danhale-git/craft/internal/docker"
+
 	"github.com/danhale-git/craft/internal/craft"
 	"github.com/spf13/cobra"
 )
@@ -23,7 +25,7 @@ The backed up world is usually a few seconds behind the world state at the time 
 		},
 		// save the world files to a backup archive
 		Run: func(cmd *cobra.Command, args []string) {
-			d := craft.NewDockerClientOrExit(args[0])
+			d := docker.NewDockerClientOrExit(args[0])
 
 			err := craft.SaveBackup(d)
 			if err != nil {
