+++ cmd/cmd.go
@@ -1,7 +1,7 @@
 package cmd
 
 import (
-	"github.com/danhale-git/craft/internal/craft"
+	"github.com/danhale-git/craft/internal/docker"
 	"github.com/spf13/cobra"
 )
 
@@ -23,7 +23,7 @@ Any number of following arguments may be provided as a mc server command.`,
 			containerName := args[0]
 			command := args[1:]
 
-			d := craft.NewDockerClientOrExit(containerName)
+			d := docker.NewDockerClientOrExit(containerName)
 
 			err := d.Command(command)
 			if err != nil {
