+++ cmd/stop.go
@@ -3,6 +3,8 @@ package cmd
 import (
 	"log"
 
+	"github.com/danhale-git/craft/internal/docker"
+
 	"github.com/danhale-git/craft/internal/craft"
 	"github.com/spf13/cobra"
 )
@@ -16,7 +18,7 @@ func init() {
 			return cobra.RangeArgs(1, 1)(cmd, args)
 		},
 		RunE: func(cmd *cobra.Command, args []string) error {
-			d := craft.NewDockerClientOrExit(args[0])
+			d := docker.NewDockerClientOrExit(args[0])
 
 			noBackup, err := cmd.Flags().GetBool("no-backup")
 			if err != nil {
