+++ internal/docker/dockerclient_test.go
@@ -1,4 +1,4 @@
-package craft
+package docker
 
 import (
 	"bufio"
